backend/requirements.txt
fastapi

uvicorn
websockets

duckdb
pydantic-settings
aiofiles

scapy

# optional but recommended
python-dotenv
paho-mqtt
httpx
paramiko
psutil

python-multipart
requests

# auth
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
zeroconf
beautifulsoup4
httpx
